Move logic from SessionHelper to ApplicationController and delete it
